// src/lib/types/enums.ts
export enum Shape {
  Tuck = "o",
  Pike = "<",
  Layout = "/",
  // Straight = '/'
  NotApplicable = "",
}

export enum Direction {
  Backward = "Backward",
  Forward = "Forward",
}

export enum InputType {
  Skill = 0,
  Routine = 1,
  RoutineSet = 2,
}

export enum Event {
  Trampoline = "TRI",
  DoubleMini = "DMT",
  Tumbling = "TUM",
}